Add files via upload
This is a very good program, it works as Online Voter Machine. It allows voter to vote for any leader. It accepts candidates name for voting purpose.  Each user can vote only once using their voter id otherwise it generates an error message in the window. At the end it displays the result of the election with the number of votes of elected candidate(winner).In this mini project, i am using list data structure to store the lists of voters who want to cast their votes and leader who wants to participate in elections.